Consolidate people and groups tabs to user / user groups shared tab
