--- frontend/src/views/Project/MembersPage/components/MemberListTab/index.tsx
@@ -1,1 +1,0 @@
-export { MemberListTab } from "./MemberListTab";
