+++ frontend/src/views/Project/MembersPage/components/MembersTab/components/index.tsx
@@ -1,0 +1,1 @@
+export { MembersSection } from "./MembersSection";
