+++ frontend/src/views/Project/MembersPage/components/MembersTab/index.tsx
@@ -1,0 +1,1 @@
+export { MembersTab } from "./MembersTab";
